backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlmodel==0.0.24
python-jose[cryptography]==3.5.0
bcrypt>=4.1.0
python-multipart==0.0.20
email-validator==2.2.0
pydantic-settings==2.10.1
